backend/ml-pipeline/requirements.txt
torch>=2.0.0
pandas>=2.0.0
supabase>=2.0.0
scikit-learn>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
